- Add warning if the Proximity version doesn't match the template version - Some better error messages - Better predicate attribute referencing - Allow setting HSB values instead of just being plus or minus - Cache plugin jars

+++ src/main/java/dev/hephaestus/proximity/xml/LayerProperty.java
@@ -22,7 +22,7 @@ public abstract class LayerProperty<T> {
 
             apply(element, "italicFont", builder::italics);
             apply(element, "capitalization", Style.Capitalization::parse, builder::capitalization);
-            apply(element, "color", Integer::decode, builder::color);
+            apply(element, "color", Long::decode, builder::color);
             apply(element, "size", Integer::parseInt, builder::size);
             apply(element, "kerning", Float::parseFloat, builder::kerning);
